Refactor procedures to use execute method
