+++ module/services/procedure/FSM/AbstractProcedure.js
@@ -2,7 +2,6 @@ import { StoreManager } from "@sveltehelpers/StoreManager.svelte";
 import { writable, derived, get, readable } from "svelte/store";
 import { localize } from "@services/utilities.js";
 import SR3ERoll from "@documents/SR3ERoll.js";
-import OpposeRollService from "@services/OpposeRollService.js";
 
 function C() {
    return CONFIG?.sr3e || {};
@@ -482,38 +481,8 @@ export default class AbstractProcedure {
       this.#disposers = [];
    }
 
-   /**
-    * Generic contested roll entry point used by Challenge.svelte
-    */
-   async challenge({ OnClose, CommitEffects } = {}) {
-      try {
-         OnClose?.();
-
-         const actor = this.caller;
-         const formula = this.buildFormula(true);
-         const baseRoll = SR3ERoll.create(formula, { actor });
-
-         await this.onChallengeWillRoll?.({ baseRoll, actor });
-
-         // Procedure-driven evaluation
-         const roll = await baseRoll.evaluate(this);
-         await baseRoll.waitForResolution();
-
-         await CommitEffects?.();
-
-         if (this.#contestIds?.length) {
-            for (const id of this.#contestIds) OpposeRollService.expireContest(id);
-            this.clearContests();
-         }
-
-         Hooks.callAll("actorSystemRecalculated", actor);
-         await this.onChallengeResolved?.({ roll, actor });
-         return roll;
-      } catch (err) {
-         DEBUG && LOG.error("Challenge flow failed", [__FILE__, __LINE__, err]);
-         ui.notifications.error(game.i18n.localize?.("sr3e.error.challengeFailed") ?? "Challenge failed");
-         throw err;
-      }
+   async execute(/* opts */) {
+      throw new Error("AbstractProcedure.execute must be overridden by subclasses");
    }
 
    /** Optional hook: run just before the roll is evaluated. Subclasses may override. */
