+++ module/services/procedure/FSM/DodgeProcedure.js
@@ -1,5 +1,6 @@
 // services/procedure/FSM/DodgeProcedure.js
 import AbstractProcedure from "@services/procedure/FSM/AbstractProcedure";
+import SR3ERoll from "@documents/SR3ERoll.js";
 import OpposeRollService from "@services/OpposeRollService.js";
 
 export default class DodgeProcedure extends AbstractProcedure {
@@ -44,18 +45,34 @@ export default class DodgeProcedure extends AbstractProcedure {
       return "Dodge the incoming attack.";
    }
 
-   // After the defender rolls, hand the result back to the service;
-   // addOpposedResponseButton is waiting on waitForResponse(contestId)
-   async onChallengeResolved({ roll /*, actor*/ }) {
+   async execute({ OnClose, CommitEffects } = {}) {
       try {
+         OnClose?.();
+
+         const actor = this.caller;
+         const formula = this.buildFormula(true);
+         const baseRoll = SR3ERoll.create(formula, { actor });
+
+         await this.onChallengeWillRoll?.({ baseRoll, actor });
+
+         const roll = await baseRoll.evaluate(this);
+         await baseRoll.waitForResolution();
+
+         await CommitEffects?.();
+
          if (!this.#contestId) {
             console.warn("[sr3e] DodgeProcedure missing contestId; cannot deliver response");
-            return;
+         } else {
+            OpposeRollService.deliverResponse(this.#contestId, roll.toJSON());
          }
-         OpposeRollService.deliverResponse(this.#contestId, roll.toJSON());
-      } catch (e) {
-         console.error("[sr3e] DodgeProcedure.onChallengeResolved failed:", e);
-         ui.notifications?.error?.("Dodge failed to report result.");
+
+         Hooks.callAll("actorSystemRecalculated", actor);
+         await this.onChallengeResolved?.({ roll, actor });
+         return roll;
+      } catch (err) {
+         DEBUG && LOG.error("Challenge flow failed", [__FILE__, __LINE__, err]);
+         ui.notifications.error(game.i18n.localize?.("sr3e.error.challengeFailed") ?? "Challenge failed");
+         throw err;
       }
    }
 }
